refactor: Move VMUtil out of struct subpackage
