+++ src/main/java/me/coley/recaf/util/ClasspathUtil.java
@@ -1,7 +1,5 @@
 package me.coley.recaf.util;
 
-import me.coley.recaf.util.struct.VMUtil;
-
 import java.io.*;
 
 import java.lang.module.ModuleFinder;
